Version 0.5.0: Add selection of format recording (WAV/MP3)

+++ app/src/main/java/com/vlad805/fmradio/service/fm/IFMRecorder.java
@@ -5,6 +5,6 @@ package com.vlad805.fmradio.service.fm;
  */
 public interface IFMRecorder {
 	void startRecord() throws RecordError;
-	void record(final byte[] data, final int length);
+	void record(final short[] data, final int length);
 	void stopRecord();
 }
